+++ package.json
@@ -12,7 +12,8 @@
     "prebuild": "node ./scripts/preInstall.js",
     "build": "npm run prebuild && GENERATE_SOURCEMAP=false react-scripts build",
     "clean": "rm -rf node_modules build",
-    "start": "react-scripts start",
+    "start": "node server/server.js",
+    "local": "react-scripts start",
     "test": "react-scripts test --watchAll=false",
     "eject": "react-scripts eject",
     "lint": "eslint '**/*.{js,ts}'",
