+++ public/index.html
@@ -5,12 +5,12 @@
   <title>Mathias Bosman • online resume</title>
   <meta name="viewport" content="initial-scale=1.0,width=device-width"/>
   <meta name="description" content="Java developer and network engineer. Currently employed at the Environment Department of the Flemish Government."/>
-  <link rel="icon" type="image/png" sizes="32x32" href="%PUBLIC_URL%/favicon-32x32.png">
-  <link rel="icon" type="image/png" sizes="16x16" href="%PUBLIC_URL%/favicon-16x16.png">
-  <link rel="shortcut icon" type="image/png" sizes="32x32" href="%PUBLIC_URL%/favicon-32x32.png">
-  <link rel="shortcut icon" type="image/png" sizes="16x16" href="%PUBLIC_URL%/favicon-16x16.png">
-  <link rel="manifest" href="%PUBLIC_URL%/site.webmanifest">
-  <link rel="apple-touch-icon" sizes="180x180" href="%PUBLIC_URL%/apple-touch-icon.png">
+  <link rel="icon" type="image/png" sizes="32x32" href="./favicon-32x32.png">
+  <link rel="icon" type="image/png" sizes="16x16" href="./favicon-16x16.png">
+  <link rel="shortcut icon" type="image/png" sizes="32x32" href="./favicon-32x32.png">
+  <link rel="shortcut icon" type="image/png" sizes="16x16" href="./favicon-16x16.png">
+  <link rel="manifest" href="./site.webmanifest">
+  <link rel="apple-touch-icon" sizes="180x180" href="./apple-touch-icon.png">
   <link rel="canonical" href="https://mathiasbosman.be">
   <meta content="dark light" name="color-scheme"/>
   <meta name="msapplication-TileColor" content="#da532c">
